[REG-FIELDS] Created text input group
